BmapCopy.py: fix a bug when copying without bmap
Fix a regression introduced in version 0.5: when copying without bmap the
script fails with the following error:

AttributeError: BmapCopy instance has no attribute '_f_bmap'

This is because we never define this attribute when there is no bmap, but still
use it. The fix is:

1. Define all attributes in the __init__ constructor.
2. Always check if _f_bmap is not None before using it.

While on it, do a minor space clean-up.

Change-Id: I3f729588d33be8e2a447fca817b994500840503a
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>