test_api_base: compare files using sha1
Instead of comparing files all the time using 'filecmp', calculate
sha1 of the image once and compare it to the copies.

This will be needed soon in order to implement compressed files testing - the
'filecmp' module does not accept file-like objects, only file names, so you
cannot feed it a compressed image.

Change-Id: I0c4db36fb89f9c5b8ba5aac116e9f25d9a17b4f6
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>