TransRead: improve user experience
This patch solves the following problem.

1. I forgot to define the proxy.
2. I run bmaptool, it blocks, and several minutes later it fails with an error
   like "Connection timed out"

I would instead like it to tell me that something is going wrong much earlier,
why should I wait for several minutes?

This patch improves the way we open URLs. Now instead of using the default
(usually long) timeout, we first try with a short timeout, and if we cannot
open the URL, we print user a warning, and then try to open with the default
timeout. The user may press Ctrl-C once he/she sees the warning, or start
checking the connectivity.

Change-Id: Ia5f387de62574d84d914c21794f782b1b41d530f
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>