Rename Fiemap.py to Filemap.py
The FIEMAP ioctl is not supported by tmpfs, so currently bmaptool fails to
create the bmap file when the file resides on tmpfs. This is unfortunate.

However, tmpfs supports 'SEEK_HOLE' which we can use instead of FIEMAP.

This patch is a preparation for adding 'SEEK_HOLE' support. Namely, we re-name
the Fiemap.py module to Filemap.py, where we'll support both FIEMAP and
SEEK_HOLE. Variables which contain 'fiemap' are also re-named so that they now
contain 'filemap' instead.

Change-Id: I20eac7adbdc3ad6d39c8ce5b67a5cbdbdf2987be
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>