Filemap: simplify the FIEMAP failure error path
Simplify the FIEMAP ioctl error handling by removing the 'looks like your
kernel does not support FIEMAP' note. I've recently got a bug report from the
field where people hit this error, and that line not was not actually appended
to the error message because the error code was something like ENOTSUPP. And
the error was anyway very clear and readable.

So let's just remove that part and simplify the code.

Change-Id: Ib098c814d70980bdcb60402eafd213c08f8c9dbd
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>