Filemap: introduce base class
We are going to introduce another class which uses the 'SEEK_HOLE' mechanism
for getting file block map, and some of the 'class Fiemap' functionality is
going to be the same in both. Let's separate that common functionality into a
separate '_FilemapBase' class which other classes will inherit.

The '_FilemapBase' class also defines the methods child classes have to
implement and documents them.

Change-Id: Ie1e05f96c1d8d9c77913b613940dc7e2e7971359
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>