TransRead: remove code duplication
We had some amount of duplicated code related to opening a tar file - introduce
a '_open_tarfile()' helper function for this instead.

On top of this, stop using 'tarfile' modules for opening 'tar.gz' and 'tar.bz2'
files since for some reasons this is a lot slower than using the
'_CompressedFile' module, like we do for 'tar.xz' files.

Change-Id: I1c35cff9056c555df7e7391da0b0213a016a9e45
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>