bmaptool: remove documentation duplication
We now have bmaptool documentation in the web, and in the man page. No need to
duplicate it in the doc string, simply because it may easily get out of date
and inconsistent with the other documentation. So simply refer to the web site
in the doc string. Also, use the docstring in the help output.

Change-Id: Idd99c668f1298633ccab12366b7838b4ce8f41f3
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>