XFS support: hint filesystem about expected maximum size of image file
XFS has feature about speculative pre-allocation of blocks based on
amount information written in the end of open file. In case of sparse
files it might lead to more mapped blocks than really expected.
Thus, truncate to desired size destination file as early as possible
and then write blocks that should be mapped the middle of it.

Closes #6

Signed-off-by: Alexander D. Kanevskiy <kad@kad.name>