port to Python3: fix UnsupportedOperation exception
Current bmaptool code assumes that if object has 'fileno'
method it can be safely called and passed to isatty function
to test if object is assiciated with a tty device.
Unfortunately this approach doesn't work in Python3 and
causes exception "UnsupportedOperation: fileno" for StringIO
objects.

Used isatty method of the file object to fix this.

Signed-off-by: Ed Bartosh <ed.bartosh@linux.intel.com>