port to Python3: use string type for SUPPORTED_BMAP_VERSION
Current code used text and int types for SUPPORTED_BMAP_VERSION
variable. Generally it's not good to mix data types like that,
but it worked in Python 2. However, it breaks in Python 3 when
trying to compare integer and string variables.

Used strings for all SUPPORTED_BMAP_VERSION variables to fix
this issue and make the code working in Python2 and Python3.

Signed-off-by: Ed Bartosh <ed.bartosh@linux.intel.com>