TransRead: handle absence of seek() better
The urllib2 file-like objects did not have 'seek()' python2 and we
emulated it by reading and discarding the data. But in python3 the
file-like objects to have the 'seek()' method, which raises an
exception:

$ python3 ./bmaptool copy https://xxx/image.raw ~/tmp/file
Traceback (most recent call last):
  File "./bmaptool", line 11, in <module>
    sys.exit(main())
  File "/home/abityuts/git/bmap-tools/bmaptools/CLI.py", line 715, in main
    args.func(args)
  File "/home/abityuts/git/bmap-tools/bmaptools/CLI.py", line 477, in copy_command
    writer.copy(False, not args.no_verify)
  File "/home/abityuts/git/bmap-tools/bmaptools/BmapCopy.py", line 584, in copy
    raise exc_info[1].with_traceback(exc_info[2])
  File "/home/abityuts/git/bmap-tools/bmaptools/BmapCopy.py", line 503, in _get_data
    self._f_image.seek(first * self.block_size)
  File "/home/abityuts/git/bmap-tools/bmaptools/TransRead.py", line 587, in seek
    self._f_objs[-1].seek(offset, whence)
io.UnsupportedOperation: seek

This patch fixes the problem by intercepting the exception and fallig-back
to emulated seek().

This patch is an improved version of a patch from Ricardo Ribalda Delgado
(ribalda in github). It basically implements what Simon McVittie (smcv in github)
suggested in a review message.

Change-Id: I783d69113a05617855e44f608b0fef3a8b22d798
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>