bmaptool: print traceback in case of failure
It is very hard to diagnos problems when there is no traceback. Lets stop hiding it and actually
print it in case of any error. First print the traceback, then the error message - this way
non-developers will first pay attention to the error message at the end, and develperes will have
useful traceback when the bug report is filed.

This patch introduces the "error_out" helper which prints the traceback, then the error message, and
then terminate the program. Here is how an artificially ingected error in the reader thread look
like without this patch:

$ bmaptool copy https://xyz/image.raw ~/tmp/file
bmaptool: ERROR: injected error

And after this patch:

$ bmaptool copy https://xyz/image.raw ~/tmp/file
Traceback (most recent call last):
  File "/home/abityuts/git/bmap-tools/bmaptools/CLI.py", line 487, in copy_command
    writer.copy(False, not args.no_verify)
  File "/home/abityuts/git/bmap-tools/bmaptools/BmapCopy.py", line 581, in copy
    reraise(exc_info[0], exc_info[1], exc_info[2])
  File "/home/abityuts/git/bmap-tools/bmaptools/BmapCopy.py", line 495, in _get_data
    raise Error("injected error")

bmaptool: ERROR: injected error

In the latter case we clearly see where the problem is.

Change-Id: Ida6a49caf1d5911b6509323ed5bf4d3afc4f6788
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>