Upade development process
The hole idea with devel is out of date and we do not use it. We use the github
process now and the project is too small for a separate 'devel' branch. We use
the github process nowadays.

Let's update the READ me file and the 'make_a_release.sh' script.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>