README: add docs from tizen.org
The Tizen IVI project did not see any visible changes for several years and it
is probably going nowhere. We do not know for how long 'tizen.org' is going to
host the old bmap-tools docs, so let's move it here.

This is a simple copypaste with little formatting effort.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>