BmapCopy: Do not show a warning if noop io scheduder is not available
New Linux Multi-Queue Block IO Queueing Mechanism does not have a noop
scheduler, which result in an error when trying to set the io scheduler.

Also according to the linux-blk maintainers:

"""
don't change the default io scheduler as none, which shouldn't
work well for slow disk, such as non-SSD.
"""

So, this patch basically does not bother the user if noop is not
available following this rationale :

If noop is not available is because the kernel is using multi queue
and then we should not change the io scheduler.

Fixes: #55
Signed-off-by: Ricardo Ribalda Delgado <ricardo.ribalda@gmail.com>