bmaptool: CLI: Fail when copying a bmap file
The current behavior is that if the user copies the bmap file
by mistake, it prints a warning but it will copy the bmap file
to the block device.

Exemple:

$ echo "test" > in
$ bmaptool create in -o in.bmap
$ bmaptool copy in.bmap out
$ cat out
<?xml version="1.0" ?>
<!-- This file contains the block map for an image file, which is basically
     a list of useful (mapped) block numbers in the image file. In other words,
...

Set the message as an error instead of a warning and update
its content to be more explicit.

Fixes f67fa08 ("bmaptool: do not fail when copying a bmap file")

Signed-off-by: Mylène Josserand <mylene.josserand@collabora.com>