Do not use subprocess pipe
We use the 'subprocess' module for running external processes, and in
few places we create sub-processes with the 'stderr=subprocess.PIPE'
argument. Howerver, we never read from the pipe, which means that it may
get filled and block the external program. This is dangerous and may
lead to deadlock situations.

This patch fixes the issue by removing the argument. If we do not read
sub-process's 'stderr', it is OK for it to inherit it from the main
program, so the error message will just go to bmaptool's standare error
stream.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>